Extract exact warning helpers partial
